+++ includes/methods_and_classes_bonus/t_interpolation_bonus.h
@@ -14,30 +14,30 @@
 # define T_INTERPOLATION_BONUS_H
 
 typedef struct s_in		t_i;
-static double			interpolate_linear(t_i inter) __attribute__((unused));
-static double			normalize(double result) __attribute__((unused));
+static float			interpolate_linear(t_i inter) __attribute__((unused));
+static float			normalize(float result) __attribute__((unused));
 
 typedef struct s_in
 {
-	double				size_delta;
-	double				min_factor;
-	double				factor_delta;
-	double				current_delta;
-	double				(*interpolate)(t_i);
-	double				(*normalize_result)(double);
+	float				size_delta;
+	float				min_factor;
+	float				factor_delta;
+	float				current_delta;
+	float				(*interpolate)(t_i);
+	float				(*normalize_result)(float);
 }						t_interpolation;
 
-static inline double	interpolate_linear(t_interpolation inter)
+static inline float	interpolate_linear(t_interpolation inter)
 {
-	double	factor;
-	double	min;
+	float	factor;
+	float	min;
 
 	min = inter.min_factor;
 	factor = min + inter.current_delta * inter.factor_delta / inter.size_delta;
 	return (factor);
 }
 
-static inline double	normalize(double result)
+static inline float	normalize(float result)
 {
 	if (result >= 2.26)
 		result = 2.26;
